changequote.js: Fix a bug on date formatting.
This bug is due to the fact that the collected date is stored as a
string and then compared to 10 as an integer to decide whether to pad it
with a leading 0; but the string is padded instead of the integer,
sometimes causing the number of the day to have two 0s (e.g. today would
be 2016-05-008 instead of 2016-05-08).