changequote.js: fix date parsing, regarding multiple spaces
The function that parses the date assumes that the string is at the
following format:

    mon, 6 aug 2018 20:19:19 +0100

However, sometimes there are multiple spaces between the string for the
day ("mon," and the number ("6"), like this:

    mon,  6 aug 2018 20:19:19 +0100

(Observed at times from emails sent with git send-email, but not all the
time. Don't ask me.)

Fix it by removing multiple spaces in date string before parsing it.