changequote.js: Replace getLocalMsgStream() with getOfflineFileStream()
Function getOfflineFileStream() was removed in Thunderbird 102 [0], in
favour of getLocalMsgStream() which is supposed to "unify local/offline
message reading".

This fixes a bug on recent Thunderbird where custom format dates would
not be applied, because the call to the now undefined
getOfflineFileStream() function would fail, CQdate would be set to "-1",
and decodeCustomizedDateSender() would never be called.

Given that getLocalMsgStream() is new, update the minimum version
requirement for the add-on.

[0] https://hg.mozilla.org/comm-central/rev/21bbb5c18ad05b92b70454648b0e5630a95e0190