Remove LegacyPrefs API (#55)
* remove LegacyPrefs API

* replace moveToStorage() by setDefaults()

The preferences were set to the default value, when there was no entry in the legacy pref branch during migration. We still need to set default prefs. I opted to check for each setting if it has a value on each startup, and set it to the default value, if not.